requirements.txt
numpy
sounddevice
scipy
python-dotenv
assemblyai
google-generativeai
